Split req listing by keyword.
