+++ ui/routes.jsx
@@ -6,6 +6,7 @@ import Keywords from './components/keywords';
 import Index from './components/index';
 import Policies from './components/policies';
 import Requirements from './components/requirements/container';
+import ReqsByKeyword from './components/requirements/by-keyword';
 import AsyncLookupSearch, { redirectIfMatched } from './components/lookup-search';
 
 export default <Router history={browserHistory} >
@@ -19,9 +20,9 @@ export default <Router history={browserHistory} >
       <IndexRoute component={Policies} />
       <Route path="search-redirect" component={AsyncLookupSearch} onEnter={redirectIfMatched} />
     </Route>
-    <Route path="requirements">
+    <Route path="requirements" component={Requirements} >
       <IndexRedirect to="./by-keyword" />
-      <Route path="by-keyword" tabName="Requirement" component={Requirements} />
+      <Route path="by-keyword" tabName="Requirement" component={ReqsByKeyword} />
     </Route>
   </Route>
 </Router>;
